fix: satisfy runtime package file-count limit for CI
Merge call-behavior inventory tests into one file and fold runtime failure codes into types.go so pkg/orchestrators/javascript/runtime stays within the 15-file maintainability limit.
