--- pkg/orchestrators/javascript/runtime/codes.go
@@ -1,12 +1,0 @@
-package workflowruntime
-
-const (
-	CodeCanceled            = "WORKFLOW_RUNTIME_CANCELED"
-	CodeTimeout             = "WORKFLOW_RUNTIME_TIMEOUT"
-	CodeScriptError         = "WORKFLOW_RUNTIME_SCRIPT_ERROR"
-	CodeUnresolvedFinal     = "WORKFLOW_RUNTIME_UNRESOLVED_FINAL"
-	CodeDeniedCapability    = "WORKFLOW_RUNTIME_DENIED_CAPABILITY"
-	CodeUnsupportedFinal    = "WORKFLOW_RUNTIME_UNSUPPORTED_FINAL"
-	CodePreExecutionInvalid = "WORKFLOW_RUNTIME_PRE_EXECUTION_INVALID"
-	CodeInvalidResult       = "WORKFLOW_RUNTIME_INVALID_RESULT"
-)
