+++ pkg/orchestrators/javascript/runtime/types.go
@@ -8,6 +8,17 @@ import (
 	"github.com/portpowered/infinite-you/pkg/orchestrators/javascript/result"
 )
 
+const (
+	CodeCanceled            = "WORKFLOW_RUNTIME_CANCELED"
+	CodeTimeout             = "WORKFLOW_RUNTIME_TIMEOUT"
+	CodeScriptError         = "WORKFLOW_RUNTIME_SCRIPT_ERROR"
+	CodeUnresolvedFinal     = "WORKFLOW_RUNTIME_UNRESOLVED_FINAL"
+	CodeDeniedCapability    = "WORKFLOW_RUNTIME_DENIED_CAPABILITY"
+	CodeUnsupportedFinal    = "WORKFLOW_RUNTIME_UNSUPPORTED_FINAL"
+	CodePreExecutionInvalid = "WORKFLOW_RUNTIME_PRE_EXECUTION_INVALID"
+	CodeInvalidResult       = "WORKFLOW_RUNTIME_INVALID_RESULT"
+)
+
 // Request carries explicit runtime inputs for one simple workflow execution.
 type Request struct {
 	Source         string
